format recipe post
