+++ src/constants/filters.js
@@ -4,9 +4,9 @@ export const MEAL_FILTERS = ["breakfast", "lunch", "dinner", "dessert"];
 
 export const FILTER_CATEGORIES = {
   course: MEAL_FILTERS,
-  serves: ["one", "two"],
-  stars: ["one", "two", "three"],
-  time: ["one", "two", "three", "four"],
+  // serves: ["one", "two"],
+  // stars: ["one", "two", "three"],
+  // time: ["one", "two", "three", "four"],
 };
 
 export const FILTERS = {
